requirements.txt
numpy==1.20.3
MDAnalysis==2.3.0
setuptools==62.3.2
gromacswrapper==0.8.2
